# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pokémon Knowledge Management Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PokemonSummary {
  id: number;
  nationalDexNumber: number;
  name: string;
  generation: number;
  types: string[];
  spriteUrl: string;
  artworkUrl: string;
  baseStatTotal: number;
  isLegendary: boolean;
  isMythical: boolean;
  isParadox?: boolean;
  isUltraBeast?: boolean;
  color?: string;
  isFavorited?: boolean;
}

export interface PokemonStats {
  hp: number;
  attack: number;
  defense: number;
  specialAttack: number;
  specialDefense: number;
  speed: number;
  total: number;
}

export interface PokemonAbility {
  name: string;
  description?: string;
  isHidden: boolean;
}

export interface TypeEffectiveness {
  doubleDamageTo?: string[];
  halfDamageTo?: string[];
  noDamageTo?: string[];
  doubleDamageFrom?: string[];
  halfDamageFrom?: string[];
  noDamageFrom?: string[];
}

export interface PokemonDetail {
  id: number;
  nationalDexNumber: number;
  name: string;
  category?: string;
  description?: string;
  generation: number;
  height?: number;
  weight?: number;
  color?: string;
  shape?: string;
  /** @nullable */
  habitat?: string | null;
  /** @nullable */
  genderRatio?: number | null;
  captureRate?: number;
  baseFriendship?: number;
  growthRate?: string;
  isLegendary?: boolean;
  isMythical?: boolean;
  isParadox?: boolean;
  isUltraBeast?: boolean;
  types: string[];
  stats: PokemonStats;
  abilities?: PokemonAbility[];
  spriteUrl: string;
  artworkUrl: string;
  /** @nullable */
  shinySpriteUrl?: string | null;
  /** @nullable */
  shinyArtworkUrl?: string | null;
  typeEffectiveness?: TypeEffectiveness;
  eggGroups?: string[];
  isFavorited?: boolean;
}

export interface PokemonListResponse {
  data: PokemonSummary[];
  total: number;
  page: number;
  limit: number;
}

export interface EvolutionDetails {
  trigger?: string;
  /** @nullable */
  minLevel?: number | null;
  /** @nullable */
  item?: string | null;
  /** @nullable */
  friendship?: number | null;
  /** @nullable */
  timeOfDay?: string | null;
  /** @nullable */
  heldItem?: string | null;
  /** @nullable */
  specialRequirement?: string | null;
}

export interface EvolutionNode {
  pokemon: PokemonSummary;
  evolvesTo?: EvolutionNode[];
  evolutionDetails?: EvolutionDetails;
}

export interface EvolutionChain {
  chain: EvolutionNode;
}

export type PokemonFormFormType = typeof PokemonFormFormType[keyof typeof PokemonFormFormType];


export const PokemonFormFormType = {
  normal: 'normal',
  shiny: 'shiny',
  regional: 'regional',
  mega: 'mega',
  gigantamax: 'gigantamax',
  alternate: 'alternate',
} as const;

export interface PokemonForm {
  id: number;
  name: string;
  formType: PokemonFormFormType;
  /** @nullable */
  region?: string | null;
  spriteUrl: string;
  artworkUrl: string;
  types?: string[];
  stats?: PokemonStats;
}

export type MoveCategory = typeof MoveCategory[keyof typeof MoveCategory];


export const MoveCategory = {
  Physical: 'Physical',
  Special: 'Special',
  Status: 'Status',
} as const;

export interface Move {
  id: number;
  name: string;
  type: string;
  category: MoveCategory;
  /** @nullable */
  power?: number | null;
  /** @nullable */
  accuracy?: number | null;
  pp?: number;
  priority?: number;
  description?: string;
}

export type PokemonMoveLearnMethod = typeof PokemonMoveLearnMethod[keyof typeof PokemonMoveLearnMethod];


export const PokemonMoveLearnMethod = {
  'level-up': 'level-up',
  tm: 'tm',
  egg: 'egg',
  tutor: 'tutor',
} as const;

export interface PokemonMove {
  move: Move;
  learnMethod: PokemonMoveLearnMethod;
  /** @nullable */
  levelLearnedAt?: number | null;
}

export interface PokemonType {
  id: number;
  name: string;
  color: string;
}

export interface PokemonTypeDetail {
  id: number;
  name: string;
  color: string;
  effectiveness: TypeEffectiveness;
  pokemonCount?: number;
}

export interface Ability {
  id: number;
  name: string;
  description: string;
  pokemonCount?: number;
}

export interface TradingCard {
  id: number;
  name: string;
  set: string;
  cardNumber: string;
  /** @nullable */
  hp?: number | null;
  rarity: string;
  /** @nullable */
  illustrator?: string | null;
  imageUrl: string;
}

export type SearchResultCategory = typeof SearchResultCategory[keyof typeof SearchResultCategory];


export const SearchResultCategory = {
  pokemon: 'pokemon',
  move: 'move',
  ability: 'ability',
} as const;

export interface SearchResult {
  id: number;
  name: string;
  category: SearchResultCategory;
  /** @nullable */
  imageUrl: string | null;
  url: string;
  /** @nullable */
  subtitle?: string | null;
}

export interface SearchResults {
  results: SearchResult[];
  total: number;
}

export interface FavoriteResult {
  success: boolean;
  isFavorited: boolean;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface ProfileUpdate {
  /** @minLength 3 */
  username?: string;
  email?: string;
  avatarUrl?: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserRole;
  favoriteCount?: number;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface MessageResponse {
  message: string;
}

export interface StatsOverview {
  totalPokemon: number;
  totalMoves: number;
  totalAbilities: number;
  totalGenerations: number;
  totalTypes: number;
  totalCards?: number;
}

export interface GenerationStat {
  generation: number;
  count: number;
  label?: string;
}

export interface TypeStat {
  type: string;
  count: number;
  color: string;
}

export type ListPokemonParams = {
page?: number;
limit?: number;
generation?: number;
type?: string;
legendary?: boolean;
mythical?: boolean;
sortBy?: ListPokemonSortBy;
sortOrder?: ListPokemonSortOrder;
};

export type ListPokemonSortBy = typeof ListPokemonSortBy[keyof typeof ListPokemonSortBy];


export const ListPokemonSortBy = {
  id: 'id',
  name: 'name',
  baseStatTotal: 'baseStatTotal',
  height: 'height',
  weight: 'weight',
} as const;

export type ListPokemonSortOrder = typeof ListPokemonSortOrder[keyof typeof ListPokemonSortOrder];


export const ListPokemonSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type ComparePokemonParams = {
ids: string[];
};

export type ListAbilitiesParams = {
page?: number;
limit?: number;
};

export type ListMovesParams = {
page?: number;
limit?: number;
type?: string;
category?: string;
};

export type SearchParams = {
q: string;
type?: SearchType;
limit?: number;
};

export type SearchType = typeof SearchType[keyof typeof SearchType];


export const SearchType = {
  pokemon: 'pokemon',
  move: 'move',
  ability: 'ability',
  all: 'all',
} as const;