# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pokémon Knowledge Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /abilities:
    get:
      operationId: listAbilities
      tags: [abilities]
      summary: List all abilities
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Abilities list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ability"
  /abilities/{id}:
    get:
      operationId: getAbility
      tags: [abilities]
      summary: Get ability details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Ability details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ability"

  # ─── Moves ──────────────────────────────────────────────────────────────────
...
  /favorites/{pokemonId}:
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add a Pokémon to favorites
      parameters:
        - name: pokemonId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteResult"
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove a Pokémon from favorites
      parameters:
        - name: pokemonId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteResult"

  # ─── Auth ───────────────────────────────────────────────────────────────────
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── Stats / Dashboard ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Ability:
      type: object
      required: [id, name, description]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        pokemonCount: { type: integer }

    # ── Trading Cards ────────────────────────────────────────────────────────
...
    FavoriteResult:
      type: object
      required: [success, isFavorited]
      properties:
        success: { type: boolean }
        isFavorited: { type: boolean }

    # ── Auth ────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string, minLength: 3 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    ProfileUpdate:
      type: object
      properties:
        username: { type: string, minLength: 3 }
        email: { type: string, format: email }
        avatarUrl: { type: string }
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token: { type: string }
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string }
        avatarUrl: { type: ["string", "null"] }
        role:
          type: string
          enum: [user, admin]
        favoriteCount: { type: integer }
        createdAt: { type: string, format: date-time }
...