# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pokémon Knowledge Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserRole;
  favoriteCount?: number;
  createdAt: Date;
}